--- formats/summary_test.md
@@ -1,78 +1,0 @@
-#### Builds
-```
-✅ build-name (build-number) 
-✅ build-name (build-number)
-✅ build-name (build-number)
-✅ build-name (build-number)
-❌ build-name (build-number): Found 2 SCA vulnerabilities (1 🎃Medium 1 👻Low)
-✅ build-name (build-number)
-✅ build-name (build-number) 
-```
-
-### Binaries
-```
-✅ binary-file-path 
-❌ build-name (build-number): Found 8 vulnerabilities
-├── 5 SCA vulnerabilities (1 💀Critical 1 🔥High 1 🎃Medium 1 👻Low 1 😐Unknown)
-└── 3 Secrets (3 🔥High)
-✅ binary-file-path
-❌ binary-file-path: Found 2 SCA vulnerabilities (1 🎃Medium 1 👻Low)
-✅ binary-file-path
-✅ binary-file-path
-```
-
-### Binaries
-```
-✅ binary-file-path 
-❌ build-name (build-number): Found 10 vulnerabilities
-├── 7 SCA vulnerabilities
-|   ├── 2 💀Critical (Not Applicable: 0)
-|   ├── 1 🔥High (Not Applicable: 1)
-|   ├── 3 🎃Medium (Not Applicable: 0)
-|   └── 1 👻Low (Not Applicable: 1)
-└── 3 Secrets (3 🔥High)
-✅ binary-file-path
-❌ binary-file-path: Found 2 SCA vulnerabilities (1 🎃Medium 1 👻Low)
-✅ binary-file-path
-✅ binary-file-path
-```
-
-### Modules
-```
-✅ module-name
-❌ module-name: Found 8 vulnerabilities
-├── 5 SCA vulnerabilities (1 💀Critical 1 🔥High 1 🎃Medium 1 👻Low 1 😐Unknown)
-└── 3 Secrets (3 🔥High)
-✅ module-path 
-❌ module-path: Found 8 vulnerabilities
-├── 5 SCA vulnerabilities (1 💀Critical 1 🔥High 1 🎃Medium 1 👻Low 1 😐Unknown)
-└── 3 Secrets (3 🔥High)
-```
-
-<br>
-
----
-
-<br>
-
-
-
-
-### Modules/Binaries/Builds (Only one)
-```
-✅ No vulnerabilities were found
-```
-
-### Modules/Binaries/Builds (Only one)
-```
-❌ Found 8 vulnerabilities
-├── 5 SCA vulnerabilities (1 💀Critical 1 🔥High 1 🎃Medium 1 👻Low 1 😐Unknown)
-├── 5 IAC vulnerabilities (1 💀Critical 1 🔥High 1 🎃Medium 1 👻Low 1 😐Unknown)
-├── 5 SAST vulnerabilities (1 💀Critical 1 🔥High 1 🎃Medium 1 👻Low 1 😐Unknown)
-└── 3 Secrets (3 🔥High)
-```
-
-### Modules/Binaries/Builds (Only one)
-```
-❌ Found 3 Secrets (3 🔥High)
-```
